smart-ats/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0
pdfminer.six==20231228
python-docx==1.1.2
reportlab==4.2.5
httpx==0.27.2
python-dotenv==1.0.1
requests==2.32.3
razorpay==1.4.2
